Move some upload code from modeSave to new coreUploader object (re: #7247)
