+++ modules/core/index.js
@@ -4,4 +4,5 @@ export { coreDifference } from './difference';
 export { coreGraph } from './graph';
 export { coreHistory } from './history';
 export { coreTree } from './tree';
+export { coreUploader } from './uploader';
 export { coreValidator } from './validator';
